display new-template prompts in cli options. create reusable helper
